Remove Network base class
