+++ dl_playground/networks/alexnet_tf.py
@@ -11,20 +11,29 @@ from tensorflow.keras.layers import (
     Conv2D, Dense, Dropout, Flatten, Input, MaxPooling2D
 )
 
-from networks.network import Network
+from utils.generic_utils import validate_config
 
 
-class AlexNet(Network):
-    """AlexNet model
-
-    `network_config` passed to the `__init__` must contain the following keys:
-    - int height: height of the input to the network
-    - int width: width of the input to the network
-    - int n_channels: number of channels of the input
-    - int n_classes: number of classes in the output layer
-    """
+class AlexNet(object):
+    """AlexNet model"""
 
     required_config_keys = {'height', 'width', 'n_channels', 'n_classes'}
+
+    def __init__(self, network_config):
+        """Init
+
+        `network_config` must contain the following keys:
+        - int height: height of the input to the network
+        - int width: width of the input to the network
+        - int n_channels: number of channels of the input
+        - int n_classes: number of classes in the output layer
+
+        :param network_config: specifies the configuration for the network
+        :type network_config: dict
+        """
+
+        validate_config(network_config, self.required_config_keys)
+        self.network_config = network_config
 
     def build(self):
         """Return the inputs and outputs to instantiate a tf.keras.Model object
